ai/requirement.txt
# Core AI and ML libraries
openai>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.8
langchain-openai>=0.0.5
sentence-transformers>=2.2.2
transformers>=4.21.0
scikit-learn>=1.3.0
numpy>=1.24.0

# PDF processing
PyPDF2>=3.0.1
pypdf>=3.17.0
reportlab>=4.0.0

# Vector database and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# HTTP and utilities
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: GPU support for faster embeddings
# torch>=2.0.0
# sentence-transformers[gpu]>=2.2.2